Cut more than 90 seconds off of integrationspecs by removing unnecessary 'not_have' capybara statements in auhtor/status_giver tests
