+++ spec/integration/author/status_giver/collecting_final_submission_files_spec.rb
@@ -15,7 +15,6 @@ RSpec.describe 'Step 5: Collecting Final Submission Files', js: true do
     context "visiting the 'Update Program Information' page" do
       it 'raises a forbidden access error' do
         visit edit_author_submission_path(submission)
-        # expect(page).to have_content 'You are not allowed to visit that page at this time, please contact your administrator'
         expect(page).to have_current_path(author_root_path)
       end
     end
@@ -23,7 +22,6 @@ RSpec.describe 'Step 5: Collecting Final Submission Files', js: true do
     context "visiting the 'Provide Committee' page" do
       it 'raises a forbidden access error' do
         visit new_author_submission_committee_members_path(submission)
-        # expect(page).to have_content 'You are not allowed to visit that page at this time, please contact your administrator'
         expect(page).to have_current_path(author_root_path)
       end
     end
@@ -31,7 +29,6 @@ RSpec.describe 'Step 5: Collecting Final Submission Files', js: true do
     context "visiting the 'Update Committee' page" do
       it 'raises a forbidden access error' do
         visit edit_author_submission_committee_members_path(submission)
-        # expect(page).to have_content 'You are not allowed to visit that page at this time, please contact your administrator'
         expect(page).to have_current_path(edit_author_submission_committee_members_path(submission))
       end
     end
@@ -39,7 +36,6 @@ RSpec.describe 'Step 5: Collecting Final Submission Files', js: true do
     context "visiting the 'Upload Format Review Files' page" do
       it 'raises a forbidden access error' do
         visit author_submission_edit_format_review_path(submission)
-        # expect(page).to have_content 'You are not allowed to visit that page at this time, please contact your administrator'
         expect(page).to have_current_path(author_root_path)
       end
     end
