+++ spec/integration/author/status_giver/collecting_format_review_files_spec.rb
@@ -17,7 +17,6 @@ RSpec.describe 'Step 3: Collecting Format Review Files', js: true do
         visit author_submissions_path
         expect(page).to have_current_path(author_submissions_path)
         expect(page).to have_content(author.last_name)
-        expect(page).to have_link('Accessibility')
       end
     end
 
@@ -32,14 +31,12 @@ RSpec.describe 'Step 3: Collecting Format Review Files', js: true do
       it "raises a forbidden access error" do
         visit new_author_submission_committee_members_path(submission)
         expect(page).to have_current_path(author_root_path)
-        expect(page).not_to have_current_path(new_author_submission_committee_members_path(submission))
       end
     end
 
     context "visiting the 'Update Committee' page" do
       it "displays the committee_members for editing" do
         visit edit_author_submission_committee_members_path(submission)
-        expect(page).not_to have_current_path(author_root_path)
         expect(page).to have_current_path(edit_author_submission_committee_members_path(submission))
       end
     end
@@ -47,7 +44,6 @@ RSpec.describe 'Step 3: Collecting Format Review Files', js: true do
     context "visiting the 'Review Committee' page" do
       it "raises a forbidden access error" do
         visit author_submission_committee_members_path(submission)
-        expect(page).not_to have_current_path(author_root_path)
         expect(page).to have_current_path(author_submission_committee_members_path(submission))
       end
     end
@@ -55,7 +51,6 @@ RSpec.describe 'Step 3: Collecting Format Review Files', js: true do
     context "visiting the 'Review Format Review Files' page" do
       it "raises a forbidden access error" do
         visit author_submission_format_review_path(submission)
-        # expect(page).to have_content 'You are not allowed to visit that page at this time, please contact your administrator'
         expect(page).to have_current_path(author_root_path)
       end
     end
@@ -63,7 +58,6 @@ RSpec.describe 'Step 3: Collecting Format Review Files', js: true do
     context "visiting the 'Upload Final Submission Files' page" do
       it "raises a forbidden access error" do
         visit author_submission_edit_final_submission_path(submission)
-        # expect(page).to have_content 'You are not allowed to visit that page at this time, please contact your administrator'
         expect(page).to have_current_path(author_root_path)
       end
     end
@@ -71,7 +65,6 @@ RSpec.describe 'Step 3: Collecting Format Review Files', js: true do
     context "visiting the 'Review Final Submission Files' page" do
       it "raises a forbidden access error" do
         visit author_submission_final_submission_path(submission)
-        # expect(page).to have_content 'You are not allowed to visit that page at this time, please contact your administrator'
         expect(page).to have_current_path(author_root_path)
       end
     end
