+++ spec/integration/author/status_giver/collecting_program_info_spec.rb
@@ -15,7 +15,6 @@ RSpec.describe 'Step 1: Collecting Program Information status', js: true do
         visit author_submissions_path
         expect(page).to have_current_path(author_submissions_path)
         expect(page).to have_content(author.last_name)
-        expect(page).to have_link('Accessibility')
       end
     end
 
@@ -23,8 +22,6 @@ RSpec.describe 'Step 1: Collecting Program Information status', js: true do
       it "raises a forbidden access error" do
         visit edit_author_submission_committee_members_path(submission)
         expect(page).to have_current_path(author_root_path)
-        # expect(page).to have_content 'You are not allowed to visit that page at this time, please contact your administrator'
-        expect(page).not_to have_current_path(edit_author_submission_committee_members_path(submission))
       end
     end
 
@@ -32,33 +29,26 @@ RSpec.describe 'Step 1: Collecting Program Information status', js: true do
       it "raises a forbidden access error" do
         visit "/author/submissions/#{submission.id}/committee_members"
         expect(page).to have_current_path(author_root_path)
-        expect(page).not_to have_current_path("author/submissions/#{submission.id}/committee_members")
-        # expect(page).to have_content('You are not allowed to visit that page at this time, please contact your administrator')
       end
     end
 
     context "visiting the 'Review Program Information' page" do
       it 'raises a forbidden access error' do
         visit author_submission_program_information_path(submission)
-        # expect(page).to have_content 'You are not allowed to visit that page at this time, please contact your administrator'
         expect(page).to have_current_path(author_root_path)
-        expect(page).not_to have_current_path(author_submission_program_information_path(submission))
       end
     end
 
     context "visiting the 'Review Committee' page" do
       it "raises a forbidden access error" do
         visit author_submission_committee_members_path(submission)
-        # expect(page).to have_content 'You have not completed the required steps to review your committee yet'
         expect(page).to have_current_path(author_root_path)
-        expect(page).not_to have_current_path(author_submission_committee_members_path(submission))
       end
     end
 
     context "visiting the 'Review Format Review Files' page" do
       it "raises a forbidden access error" do
         visit author_submission_format_review_path(submission)
-        # expect(page).to have_content 'You are not allowed to visit that page at this time, please contact your administrator'
         expect(page).to have_current_path(author_root_path)
       end
     end
@@ -66,7 +56,6 @@ RSpec.describe 'Step 1: Collecting Program Information status', js: true do
     context "visiting the 'Upload Final Submission Files' page" do
       it "raises a forbidden access error" do
         visit author_submission_edit_final_submission_path(submission)
-        # expect(page).to have_content 'You are not allowed to visit that page at this time, please contact your administrator'
         expect(page).to have_current_path(author_root_path)
       end
     end
@@ -74,7 +63,6 @@ RSpec.describe 'Step 1: Collecting Program Information status', js: true do
     context "visiting the 'Review Final Submission Files' page" do
       it "raises a forbidden access error" do
         visit author_submission_final_submission_path(submission)
-        # expect(page).to have_content 'You are not allowed to visit that page at this time, please contact your administrator'
         expect(page).to have_current_path(author_root_path)
       end
     end
