+++ spec/integration/author/status_giver/waiting_for_final_submission_response_spec.rb
@@ -13,7 +13,6 @@ RSpec.describe "Step 7: Waiting for Final Submission Response'", js: true do
     context "visiting the 'Update Program Information' page" do
       it "raises a forbidden access error" do
         visit edit_author_submission_path(submission)
-        # expect(page).to have_content 'You are not allowed to visit that page at this time, please contact your administrator'
         expect(page).to have_current_path(author_root_path)
       end
     end
@@ -21,7 +20,6 @@ RSpec.describe "Step 7: Waiting for Final Submission Response'", js: true do
     context "visiting the 'Provide Committee' page" do
       it "raises a forbidden access error" do
         visit new_author_submission_committee_members_path(submission)
-        # expect(page).to have_content 'You are not allowed to visit that page at this time, please contact your administrator'
         expect(page).to have_current_path(author_root_path)
       end
     end
@@ -29,7 +27,6 @@ RSpec.describe "Step 7: Waiting for Final Submission Response'", js: true do
     context "visiting the 'Update Committee' page" do
       it "raises a forbidden access error" do
         visit edit_author_submission_committee_members_path(submission)
-        # expect(page).to have_content 'You are not allowed to visit that page at this time, please contact your administrator'
         expect(page).to have_current_path(author_root_path)
       end
     end
@@ -37,7 +34,6 @@ RSpec.describe "Step 7: Waiting for Final Submission Response'", js: true do
     context "visiting the 'Upload Format Review Files' page" do
       it "raises a forbidden access error" do
         visit author_submission_edit_format_review_path(submission)
-        # expect(page).to have_content 'You are not allowed to visit that page at this time, please contact your administrator'
         expect(page).to have_current_path(author_root_path)
       end
     end
@@ -66,7 +62,6 @@ RSpec.describe "Step 7: Waiting for Final Submission Response'", js: true do
     context "visiting the 'Upload Final Submission Files' page" do
       it "raises a forbidden access error" do
         visit author_submission_edit_final_submission_path(submission)
-        # expect(page).to have_content 'You are not allowed to visit that page at this time, please contact your administrator'
         expect(page).to have_current_path(author_root_path)
       end
     end
