+++ spec/integration/author/status_giver/waiting_for_publication_release_spec.rb
@@ -19,14 +19,12 @@ RSpec.describe "Last Step: 'waiting for publication release'", js: true do
         visit author_submissions_path
         expect(page).to have_current_path(author_submissions_path)
         expect(page).to have_content(author.last_name)
-        expect(page).to have_link('Accessibility')
       end
     end
 
     context "visiting the 'Update Program Information' page" do
       it "raises a forbidden access error" do
         visit edit_author_submission_path(submission)
-        # expect(page).to have_content 'You are not allowed to visit that page at this time, please contact your administrator'
         expect(page).to have_current_path(author_root_path)
       end
     end
@@ -34,7 +32,6 @@ RSpec.describe "Last Step: 'waiting for publication release'", js: true do
     context "visiting the 'Provide Committee' page" do
       it "raises a forbidden access error" do
         visit new_author_submission_committee_members_path(submission)
-        # expect(page).to have_content 'You are not allowed to visit that page at this time, please contact your administrator'
         expect(page).to have_current_path(author_root_path)
       end
     end
@@ -43,7 +40,6 @@ RSpec.describe "Last Step: 'waiting for publication release'", js: true do
       before { visit edit_author_submission_committee_members_path(submission) }
 
       it "raises a forbidden access error" do
-        # expect(page).to have_content 'You are not allowed to visit that page at this time, please contact your administrator'
         expect(page).to have_current_path(author_root_path)
       end
     end
@@ -51,7 +47,6 @@ RSpec.describe "Last Step: 'waiting for publication release'", js: true do
     context "visiting the 'Upload Format Review Files' page" do
       it "raises a forbidden access error" do
         visit author_submission_edit_format_review_path(submission)
-        # expect(page).to have_content 'You are not allowed to visit that page at this time, please contact your administrator'
         expect(page).to have_current_path(author_root_path)
       end
     end
@@ -80,7 +75,6 @@ RSpec.describe "Last Step: 'waiting for publication release'", js: true do
     context "visiting the 'Upload Final Submission Files' page" do
       it "raises a forbidden access error" do
         visit author_submission_edit_final_submission_path(submission)
-        # expect(page).to have_content 'You are not allowed to visit that page at this time, please contact your administrator'
         expect(page).to have_current_path(author_root_path)
       end
     end
